% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adapt}
\alias{adapt}
\title{Data set used in the package vignette}
\format{
A data frame with 180 rows and 5 variables
\describe{
  \item{y}{the outcome}
  \item{x}{a variable associated with y and treatment}
  \item{source}{character variable with 3 levels: "Primary", "Supp1", "Supp2"}
  \item{treatment}{treatment variable. 1 = treated, 0 = untreated}
  \item{compliant}{compliance indicator. 1 = compliant to treatment,
  0 = noncompliant to treatment}
}
}
\usage{
adapt
}
\description{
A simulated data set used in the package vignette. Data generating
mechanism is adapted from Hill (2011). Includes 3 data sources to illustrate
borrowing information in estimate the population average treatment effect.
}
\examples{
data(adapt)
head(adapt)
}
\references{
Hill, Jennifer L. (2011) Bayesian Nonparametric Modeling for Causal Inference.
Journal of Computational and Graphical Statistics, 20:1, 217-240.
}
\keyword{datasets}
