% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pate.R
\name{pate}
\alias{pate}
\title{Population Average Treatment Effect (PATE)}
\usage{
pate(formula, estimator = c("BART", "bayesian_lm"), data, src_var,
  primary_source, trt_var, subset, na.action, ndpost = 1000, beta = 2,
  gf = 16, eb = FALSE, prior_calibration = c("none", "empirical"),
  ...)
}
\arguments{
\item{formula}{An object of class formula. The left hand side must be the outcome,
and the right hand side must include the treatment variable. To adjust for confounding,
the right hand side must also include the confounders. For the Bayesian linear model,
the use is responsible for specifying the function form.}

\item{estimator}{"bayesian_lm" or "BART". If "bayesian_lm", a Bayesian linear model with
a normal inverse-gamma prior. If "BART", Bayesian Additive Regression Trees}

\item{data}{the data frame with the outcome variable and all variables in the formula}

\item{src_var}{a character variable which variable indicates the source variable.
Must match a column name in the data.}

\item{primary_source}{character variable indicating the primary source. Must match
one of the values of \code{src_var}.}

\item{trt_var}{which variable indicates the treatment.
Must match a column name in the data. Must be coded as numeric values 0 and 1, 0 for
untreated, 1 for treated.}

\item{subset}{an expression for subsetting the data}

\item{na.action}{how to handle missing values. Will be eliminated as an argument.}

\item{ndpost}{number of draws from the posterior}

\item{beta}{prior parameter on controlling the depth of the trees on the BART prior}

\item{gf}{paramater control the scalar multiplying sigma squared. More details needed here.}

\item{eb}{empirical bayes estimator for gf? Will be eliminated as an argument.}

\item{prior_calibration}{how should the prior on regression parameters be specified for
bayesian linear model? Will be eliminated as an argument.}

\item{...}{additional arguments passed to BART}
}
\description{
Estimates the population average treatment effect from a primary data source
with potential borrowing from supplemental sources. Adjust for confounding by fitting
a conditional mean model with the treatment variable and confounding variables.
}
\details{
To address confounding, the combr package estimates the PATE by fiting a
model for the conditional mean given treatment and confounders. Currently,
two models are available, a Bayesian linear model with an inverse-gamma prior,
and Bayesian Additive Regression Trees (BART; Chipman & McCulloch, 2010).
The user must specify a
formula for the conditional mean. This requires more thought for the
Bayesian linear model as the analyst must carefully consider the functional
form of the regression relationship. For BART, the right hand side of the
formula need only include the confounders and the treatment variable without
specification of the functional form. If there is no confounding, the right hand
side of the formula needs to include the treatment variable only.

Borrowing between data sources is done with
Multisource Exchangeability Models
(MEMs; Kaizer et al., 2018) . MEMs borrow by assuming that each supplementary data source
is either "exchangable", or not, with the primary data source. Two data sources
are considered exchangeable if their model parameters are equal. Each data source
can be exchangeable with the primary data, or not, so if there are $r$ data
sources, there are $2 ^ r$ possible configurations regarding the exchangeability
assumptions. Each of these configurations corresponds to a single MEM.
The parameters for each MEM are estimated, and we compute a posterior probability
for each. The posterior density of the PATE is a weighted posterior across all
possible MEMs.
}
\examples{
data(adapt)

est <- pate(y ~ treatment*x + treatment*I(x ^ 2), data = adapt,
 estimator = "bayesian_lm", src_var = "source", primary_source = "Primary",
 trt_var = "treatment")

est

}
\references{
Chipman, H. & McCulloch, R. (2010) BART: Bayesian additive regression trees. Annals
of Applied Statistics, 4(1): 266-298.

Kaizer, Alexander M., Koopmeiners, Joseph S., Hobbs, Brian P. (2018) Bayesian
hierarchical modeling based on multisource exchangeability. Biostatistics,
19(2): 169-184.
}
