% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pate}
\alias{print.pate}
\title{Printing Population Average Treatment Effects Object}
\usage{
\method{print}{pate}(x, ...)
}
\arguments{
\item{x}{an object of class "\code{pate}"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} method for class "\code{pate}".
}
