% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.pate}
\alias{summary.pate}
\title{Summary of Population Average Treatment Effect}
\usage{
\method{summary}{pate}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{pate}".}

\item{...}{further arguments passed to or from other methods/}
}
\description{
\code{summary} for class "\code{pate}".
}
